pyproject.toml
[project]
name = "ai-agent-project"
version = "0.1.0"
description = "AI agent specialized in stock information and recommendations"
requires-python = ">=3.10"
dependencies = [
    "langchain",
    "langchain-upstage",
    "yfinance",
    "streamlit",
    "python-dotenv",
    "langgraph>=1.0.1",
    "deep-translator>=1.11.4",
    "tavily>=1.0.0",
    "langchain-community>=0.3.31",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "chromadb>=1.3.0",
    "langchain-chroma>=0.2.6",
    "langchain-huggingface>=0.3.1",
    "langchain-text-splitters>=0.3.11",
    "pypdf>=4.3.1",
    "reportlab>=4.4.4",
    "pillow>=12.0.0",
    "numpy>=2.2.6",
    "sentence-transformers>=5.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest",
    "black",
    "ruff",
]
